requirements.txt
# RSS Feed Processing
feedparser>=6.0.0
requests>=2.31.0

# Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# AI Processing
anthropic>=0.40.0

# HTML/Content
beautifulsoup4>=4.12.0
html2text>=2024.2.26

# Utilities
python-dateutil>=2.8.0
openpyxl>=3.1.0

# Social Media
atproto>=0.0.50      # Bluesky/AT Protocol
